refactored freight
