+++ src/content.js
@@ -29,7 +29,6 @@ import { getParam } from "./material-utils";
 import { extractEmissionsFactor } from "./material-utils";
 import { getQuestionLCA } from "./material-utils";
 import {
-  getFreightData,
   hideAndClearMasterContainer,
   hidePopup,
   setupLCABannerAndFloatingMenu,
@@ -38,6 +37,7 @@ import {
   formatToSignificantFigures,
   getBeefInfo
 } from "./popup-content";
+import { getFreightData } from "./autodetect/freight/freight-utils";
 import { handleCO2eEquivalencyChange } from "./utils/ui-utils";
 import { injectPopupContent } from "./popup-content";
 import { updateFreightContent } from "./popup-content";
