+++ src/popup-content.js
@@ -10,6 +10,7 @@ import { getRecommendedModels } from "./autodetect/phone/phone-utils";
 import { getFedexDataChange, observeFedexShippingOptions,
   recordAllInputChange, recordPackageTypeChange, recordFromToAddressChange } from "./autodetect/freight/freight-tracker";
 import { getFreightHTMLContent } from "./autodetect/freight/freight-ui";
+import { getFreightData } from "./autodetect/freight/freight-utils";
 import { handleCO2eEquivalencyChange, hideElement, showElement } from "./utils/ui-utils";
 import { getPhoneEmissionsSkeleton, displayPhoneSpecEmissions, displaySideBySideComparison } from "./autodetect/phone/phone-ui";
 
@@ -629,8 +630,6 @@ export async function injectPopupContent(
   if (popupCase === "phone") {
     const phoneSkeletonHTML = getPhoneEmissionsSkeleton();
     masterContainer.insertAdjacentHTML("beforeend", phoneSkeletonHTML);
-    // Stop input propogation of the search-phone input. Some websites have their own input behavior. We need to disable them.
-    // stopInputPropogation(shadowRoot.getElementById("search-phone"));
     // Delay the execution of showPhoneEmissions to ensure DOM elements are available
     setTimeout(async () => {
       masterContainer.classList.remove("lca-viz-hidden");
@@ -654,19 +653,6 @@ export async function injectPopupContent(
     }, 0);
   }
 }
-
-// Stops the input propogation of the input element
-// function stopInputPropogation(input) {
-//   input.addEventListener("keydown", (event) => {
-//     event.stopPropagation();
-//   });
-//   input.addEventListener("keyup", (event) => {
-//     event.stopPropagation();
-//   });
-//   input.addEventListener("keypress", (event) => {
-//     event.stopPropagation();
-//   });
-// }
 
 // Returns the HTML code for the floating menu
 function getLCAFloatingMenu() {
@@ -1038,49 +1024,8 @@ function injectGoogleMaps(mode) {
   mapsContainer.appendChild(iframe);
 }
 
-// Handles searching for a phone model from the database
-// function handlePhoneSearch() {
-//   const searchInput = shadowRoot.getElementById("search-phone");
-//   searchInput.addEventListener("keyup", search);
-// }
-
-// // Searches the phone model from the database
-// function search() {
-//   let input = shadowRoot.getElementById("search-phone");
-//   let filter = input.value.toUpperCase();
-//   let container = shadowRoot.querySelector(".phone-model-container");
-//   const children = container.children;
-//   for (let i = 0; i < children.length; i++) {
-//     let textValue = children[i].textContent || children[i].innerText;
-//     if (textValue.toUpperCase().indexOf(filter) > -1) {
-//       children[i].style.display = "";
-//     } else {
-//       children[i].style.display = "none";
-//     }
-//   }
-// }
-
 // Handles the interaction of the phone comparison, including the compare and close button
 async function handlePhoneCompare() {
-  // const compareBtn = shadowRoot.querySelector(".compare-btn");
-  // const comparePhone = shadowRoot.querySelector(".compare-phone");
-  // const compareContainer = shadowRoot.querySelector(".compare-container");
-  // const slideContent = shadowRoot.querySelector(".slide-content");
-
-  // compareBtn.addEventListener("click", async () => {
-  //   // await populatePhoneModel();
-  //   comparePhone.classList.add("down");
-  //   compareContainer.classList.add("hide");
-  //   slideContent.classList.add("slide-down");
-  //   handleSideBySideSelection();
-  // });
-
-  // const closeBtn = shadowRoot.querySelector(".close-phone-btn");
-  // closeBtn.addEventListener("click", () => {
-  //   comparePhone.classList.remove("down");
-  //   compareContainer.classList.remove("hide");
-  //   slideContent.classList.remove("slide-down");
-  // });
   await populatePhoneModel();
 
   // For the new 2 recommended phone models
@@ -1098,78 +1043,6 @@ async function handlePhoneCompare() {
     });
   });
 }
-
-// Handles the selection of a phone model in the list
-// function handleSideBySideSelection() {
-//   const phoneNodeList = shadowRoot.querySelector(
-//     ".phone-model-container"
-//   ).children;
-//   Array.from(phoneNodeList).forEach((phone) => {
-//     phone.addEventListener("click", (event) => {
-//       const phoneId = parseInt(event.target.id);
-
-//       const comparePhone = shadowRoot.querySelector(".compare-phone");
-//       const compareContainer = shadowRoot.querySelector(".compare-container");
-//       const slideContent = shadowRoot.querySelector(".slide-content");
-//       comparePhone.classList.remove("down");
-//       compareContainer.classList.remove("hide");
-//       slideContent.classList.remove("slide-down");
-
-//       displaySideBySideComparison(phoneId);
-//     });
-//   });
-// }
-
-// Handles the changing of different reference units for phone emissions flow.
-// (i.e. changing between "~ kg of trash burned", "~ of miles driven", and "~ of trees cut down" every 3 seconds)
-// export function handleCO2eEquivalencyChange(isRawMaterial = false) {
-//   let selector = shadowRoot;
-//   if (isRawMaterial) {
-//     selector = document;
-//   }
-//   const unitSelect = selector.getElementById("lca-viz-unit-select");
-//   const unitDivsContainer = selector.querySelectorAll(
-//     ".lca-viz-unit-container"
-//   );
-//   // Initialize: show the first unit-div by default
-//   let currentIndex = 0;
-//   if (unitSelect && unitDivsContainer) {
-//     console.log("CALLING handleCO2EquivalencyChange");
-//     unitDivsContainer.forEach((container) => {
-//       container.children[currentIndex].classList.add("lca-viz-show");
-//     });
-
-//     unitSelect.addEventListener("change", (e) => {
-//       const selectedIndex = parseInt(e.target.value);
-//       console.log("selectedIndex = " + selectedIndex);
-//       showSelectedUnit(selectedIndex);
-//     });
-//   } else {
-//     console.log("handleCO2EquivalencyChange CANNOT be called");
-//   }
-
-//   // Function to change the displayed unit-div based on dropdown selection
-//   function showSelectedUnit(index) {
-//     unitDivsContainer.forEach((container) => {
-//       console.log("currentIndex = " + currentIndex);
-//       console.log("newIndex = " + index);
-//       const oldUnitDiv = container.children[currentIndex];
-//       const newUnitDiv = container.children[index];
-//       console.log("unitDivs = ");
-//       console.dir(oldUnitDiv);
-//       // Hide the current unit-div
-//       oldUnitDiv.classList.remove("lca-viz-show");
-//       oldUnitDiv.classList.add("lca-viz-hide");
-
-//       // After fade-out, remove the hide class and show the selected unit
-//       setTimeout(() => {
-//         oldUnitDiv.classList.remove("lca-viz-hide");
-//         newUnitDiv.classList.add("lca-viz-show");
-//         currentIndex = index;
-//       }, 300);
-//     });
-//   }
-// }
 
 /**
  * Populates the phone model that can be used for side-by-side emissions comparison
@@ -1191,112 +1064,6 @@ async function populatePhoneModel() {
     phoneCompetitorContainer.innerHTML += phoneElement;
   });
 }
-
-// Displays the carbon emission of the phone being analyzed in the web page.
-// function displayPhoneSpecEmissions() {
-//   const data = currentPhoneData;
-//   console.log("phone data = ");
-//   console.log(data);
-
-//   const container = shadowRoot.querySelector(".phone-spec-container");
-//   const specs = data.specs;
-//   const deviceName = data.device;
-
-//   container.innerHTML += `
-//     <div class="flex-stretch lca-viz-title-and-question lcz-mt-8">
-//       <p class="phone-spec-title" id="currentPhone"><b>${deviceName} Estimated Carbon Emissions</b></p>
-//       <div class="btn lca-viz-btn-primary lca-viz-tooltip"><img src="${question_icon}" alt="Hover me to get additional information" class="lcz-icon-20" id="lca-viz-q-icon">
-//         <div class="left">
-//           <h3 class="fz-12 lca-lexend">How are phone emissions calculated?</h3>
-//           <p class="fz-12 lca-lexend">We use data from phone companies' product carbon footprint reports. If there is no data, a large language model (LLM) is used to estimate emissions based on publicly available data online.</p>
-//           <i></i>
-//         </div>
-//       </div>
-//     </div>
-//     <div class="flex-center cg-8 fz-16">
-//       <p>CO2e Equivalency: </p>
-//       <select id="lca-viz-unit-select" class="lcz-br-4 pd-4">
-//         <option value="0">Miles driven 🚗</option>
-//         <option value="1">Trees offset 🌳</option>
-//         <option value="2">Beef Consumed 🥩</option>
-//       </select>
-//     </div>
-//   `;
-
-//   const compareWith = shadowRoot.getElementById("lca-viz-compare-with");
-//   compareWith.innerHTML = deviceName;
-
-//   // let mostGreenOption = footprints[0];
-//   let mostGreenOption = specs[0];
-//   specs.forEach((spec, index) => {
-//     if (index !== 0) {
-//       const co2eValue = parseFloat(spec.co2e);
-//       const mostGreenCo2eValue = parseFloat(mostGreenOption.co2e);
-//       if (co2eValue < mostGreenCo2eValue) {
-//         mostGreenOption = spec;
-//       }
-//     }
-//   });
-
-//   specs.forEach((spec, index) => {
-//     const co2eValue = parseFloat(spec.co2e);
-//     const mostGreenCo2eValue = parseFloat(mostGreenOption.co2e);
-//     const percentageIncrease =
-//       ((co2eValue - mostGreenCo2eValue) / mostGreenCo2eValue) * 100;
-
-//     const isMostGreen = spec.storage === mostGreenOption.storage;
-//     container.innerHTML += `
-//       <div class="lcz-details-container fz-16" id=${index + "-c"}>
-//         <div class="flex-center ${isMostGreen ? "most-green" : ""} cg-4">
-//           <p><b>${spec.storage} </b>&nbsp;</p>
-//           ${
-//             isMostGreen
-//               ? `<img src="${most_green_icon}" class="lcz-icon-16 emissions-diff-minus lcz-br-4 lcz-margin-0 lca-viz-MEF" title="This is the most eco-friendly option" alt="Most eco-friendly option">`
-//               : `<span class="emissions-diff-plus fz-12 lcz-br-4 lcz-margin-0"><b>(+${percentageIncrease.toFixed(
-//                   0
-//                 )}% emissions)</b></span>`
-//           }
-//         </div>
-//         <div class="flex-center co2e-data-container pd-8 lcz-br-8 cg-4 lca-viz-lexend-reg">
-//           <p class="lcz-margin-0">${co2eValue} kg CO2e</p>
-//           <img src="${equivalent_icon}" class="lcz-icon-16" alt="Equivalent to">
-//           <div class="lca-viz-unit-container phone flex-center cg-4">
-
-//             <div class="lca-viz-unit-div">
-//               <div class="flex-center lca-viz-justify-center cg-8">
-//                 <p class="lcz-margin-0 lcz-grey-text fz-16 lca-viz-text-align-center">${Math.ceil(
-//                   co2eValue * 2.5
-//                 )} miles driven by a car &nbsp;🚗</p>
-//               </div>
-//             </div>
-
-//             <div class="lca-viz-unit-div">
-//               <div class="flex-center lca-viz-justify-center cg-8">
-//                 <p class="lcz-margin-0 lcz-grey-text fz-16 lca-viz-text-align-center">${(
-//                   co2eValue * 0.048
-//                 ).toFixed(1)} trees annually &nbsp;🌳</p>
-//               </div>
-//             </div>
-
-//             <div class="lca-viz-unit-div">
-//               <div class="flex-center lca-viz-justify-center cg-8">
-//                 <p class="lcz-margin-0 lcz-grey-text fz-16 lca-viz-text-align-center">${(
-//                   co2eValue * 0.033
-//                 ).toFixed(2)} kg of beef consumed &nbsp;🥩</p>
-//               </div>
-//             </div>
-
-//           </div>
-//         </div>
-//       </div>
-//       `;
-//   });
-
-//   const dataSource = getDataSource(currentPhoneData);
-//   container.innerHTML += dataSource;
-
-//   handleCO2eEquivalencyChange(shadowRoot);
-// }
 
 // Use this function to display a loading animation while waiting for the API calls
 export async function hideLoadingIcon(boxNumber = "") {
@@ -1341,311 +1108,6 @@ function formatRegionNames(input) {
   return regionName;
 }
 
-/**
- * Determines the appropriate transportation mode based on the shipping type and locations
- * @param {String} shippingType The Fedex shipping type (e.g. "fedex ground", "fedex 1day freight")
- * @param {String} fromValue The starting location
- * @param {String} toValue The destination location
- * @returns {String} The appropriate transportation mode, either "air" or "ground"
- */
-async function getFedexTransportMode(shippingType, fromValue, toValue) {
-  const shipping_modes = {
-    "fedex express saver": "ground",
-    "fedex ground": "ground",
-    "fedex home delivery": "ground",
-    "fedex freight priority": "ground",
-    "fedex freight economy": "ground",
-    "fedex sameday freight": "air",
-    "fedex regional economy": "ground",
-    "fedex 1day freight": "air",
-    "fedex 2day freight": "air",
-    "fedex 3day freight": "air",
-    "fedex international priority express": "air",
-    "fedex international first": "air",
-    "fedex international next flight": "air",
-    "fedex international priority": "air",
-    "fedex international economy": "air",
-    "fedex international connect plus": "air",
-    "fedex international priority freight": "air",
-    "fedex international deferred freight": "air",
-    "fedex international economy freight": "air",
-    "fedex international ground": "air",
-  };
 
-  let mode = shipping_modes[shippingType];
-  if (mode) {
-    return mode;
-  } else {
-    // If it takes longer than this amount of hours by road, then assume the shipping type is air
-    let hoursThreshold;
-    if (
-      shippingType === "fedex sameday" ||
-      shippingType === "fedex priority overnight" ||
-      shippingType === "fedex priority" ||
-      shippingType === "fedex priority express" ||
-      shippingType === "fedex standard overnight" ||
-      shippingType === "fedex first overnight" ||
-      shippingType === "fedex first overnight freight"
-    ) {
-      hoursThreshold = 6;
-    } else if (shippingType === "fedex 2day am") {
-      hoursThreshold = 12;
-    } else if (shippingType === "fedex 2day") {
-      hoursThreshold = 18;
-    } else {
-      return null;
-    }
-    try {
-      const response = await fetch(LCA_SERVER_URL + "/api/travel-time", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          from: fromValue,
-          to: toValue,
-        }),
-      });
-      const responseData = await response.json();
-      // travelTime is in seconds by default
-      const travelTime = responseData.rows[0].elements[0].duration.value;
-      if (travelTime / 3600 < hoursThreshold) {
-        return "ground";
-      }
-      return "air";
-    } catch (error) {
-      console.error("Error in getFedexTransportMode:", error);
-      return null;
-    }
-  }
-}
 
-/**
- * Injects a banner into the fedex webpage to highlight the most eco-friendly shipping options in the given array.
- * @param {Array} optionsArray Array containing a list of shipping options
- */
-function showGreenestOption(optionsArray) {
-  const availableOptions = document.querySelectorAll(
-    ".fdx-c-definitionlist__description--small"
-  );
-  availableOptions.forEach((option) => {
-    const formattedOption = option.outerText.toLowerCase().replace(/®/g, "");
-    if (optionsArray.includes(formattedOption)) {
-      const parentNode = option.parentNode.parentNode.parentNode.parentNode;
-      const priceButton = parentNode.querySelector(".magr-c-rates__button");
 
-      const newContainerHTML = ` <div class="lca-viz-greenest-shipping lcz-mb-16"> ${priceButton.outerHTML} <div class="flex-center lcz-br-4 pd-8 cg-8 green-shipping lca-viz-justify-center"> <img src="${lca_48}" alt="Most eco friendly" class="lcz-icon-16"> <span>Most eco-friendly</span> </div> </div> `;
-      // Replace the original button with the new container
-      priceButton.outerHTML = newContainerHTML;
-    }
-  });
-}
-
-/**
- * Gets the freight emissions from the Climatiq API
- * @param {Object} data The data to be sent to the Climatiq API
- * @returns The freight emissions from the Climatiq API
- */
-async function getFreightEmissions(data) {
-  try {
-    const response = await fetch(LCA_SERVER_URL + "/api/freight", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(data),
-    });
-
-    if (!response.ok) {
-      console.log(`HTTP error! Status: ${response.status}`);
-      return null;
-    }
-
-    const responseData = await response.json();
-    console.log("API Response: ", responseData);
-    return responseData;
-  } catch (error) {
-    console.error("Error:", error);
-  }
-}
-
-/**
- * Returns the freight data used to display the freight emissions and geo-map.
- * @param {boolean} isHighlight default is false. isHighlight is true when this function is called from the LCA brush scenario.
- * @returns The freight data used to display the freight emissions and geo-map.
- */
-export async function getFreightData(
-  fromAddress,
-  toAddress,
-  totalWeight,
-  currShippingOptions,
-  isHighlight = false
-) {
-  let groundMode, airMode;
-  if (!isHighlight) {
-    ({ groundMode, airMode } = await categorizeShippingOption(
-      fromAddress,
-      toAddress,
-      currShippingOptions
-    ));
-  } else {
-    groundMode = [""];
-    airMode = [""];
-  }
-
-  let freightGroundData;
-  let freightAirData;
-  let aData;
-  let gData;
-
-  if (groundMode.length > 0) {
-    const groundData = formatFreightData(
-      fromAddress,
-      toAddress,
-      "ground",
-      totalWeight
-    );
-    freightGroundData = await getFreightEmissions(groundData);
-    if (freightGroundData) {
-      gData = {
-        co2eValue: parseFloat(freightGroundData.co2e.toFixed(2)),
-        groundMode: groundMode,
-      };
-    } else {
-      gData = null;
-    }
-  }
-  if (airMode.length > 0) {
-    const airData = formatFreightData(
-      fromAddress,
-      toAddress,
-      "air",
-      totalWeight
-    );
-    freightAirData = await getFreightEmissions(airData);
-    if (freightAirData) {
-      aData = {
-        co2eValue: parseFloat(freightAirData.co2e.toFixed(2)),
-        airMode: airMode,
-      };
-    } else {
-      aData = null;
-    }
-  }
-
-  const formattedFreightData = {
-    from: fromAddress,
-    to: toAddress,
-    air: aData,
-    ground: gData,
-  };
-
-  const freightData = {
-    formatted: formattedFreightData,
-    originalAir: freightAirData,
-    originalGround: freightGroundData,
-  };
-  return freightData;
-}
-
-/**
- * Categorizes the shipping options into ground and air modes
- * @param {String} fromAddress The origin address
- * @param {String} toAddress The destination address
- * @param {Array} shippingOptions The list of all given Fedex shipping options
- * @returns Two arrays, one containing the shipping options that have ground transport mode,
- *          another containing the shipping options that have air transport mode.
- */
-async function categorizeShippingOption(
-  fromAddress,
-  toAddress,
-  shippingOptions
-) {
-  let airMode = [];
-  let groundMode = [];
-  // Use Promise.all to wait for all async operations to complete
-  const modes = await Promise.all(
-    shippingOptions.map(async (option) => {
-      const mode = await getFedexTransportMode(option, fromAddress, toAddress);
-      // console.log('option: ', option, '\nmode: ', mode);
-      return { option, mode };
-    })
-  );
-
-  // Categorize the options based on the mode
-  modes.forEach(({ option, mode }) => {
-    if (mode === "air") {
-      // console.log('pushed to air');
-      airMode.push(option);
-    }
-    if (mode === "ground") {
-      // console.log('pushed to ground');
-      groundMode.push(option);
-    }
-  });
-
-  showGreenestOption(groundMode);
-  return { groundMode, airMode };
-}
-
-/**
- * Formats the freight data for the Climatiq API
- * @param {String} fromAddress The origin address
- * @param {String} toAddress The destination address
- * @param {String} mode The mode of transportation
- * @param {Number} totalWeight The total weight of the cargo
- * @returns The formatted freight data
- */
-function formatFreightData(fromAddress, toAddress, mode, totalWeight) {
-  let transportMode;
-  if (mode === "air") {
-    transportMode = [
-      {
-        transport_mode: "road",
-      },
-      {
-        transport_mode: "air",
-      },
-      {
-        transport_mode: "road",
-      },
-    ];
-  }
-  if (mode === "ground") {
-    transportMode = [
-      {
-        transport_mode: "road",
-      },
-    ];
-  }
-  let rFrom = [
-    {
-      location: {
-        query: fromAddress,
-      },
-    },
-  ];
-  let rTo = [
-    {
-      location: {
-        query: toAddress,
-      },
-    },
-  ];
-  let cargo = [
-    {
-      weight: totalWeight,
-      weight_unit: "kg",
-    },
-  ];
-
-  // Combine the route components into one array
-  let route = rFrom.concat(transportMode).concat(rTo);
-  // Create the final data object
-  const data = {
-    route: route,
-    cargo: cargo[0], // Access the first (and only) object in the cargo array
-  };
-  // console.log(JSON.stringify(data, null, 2));
-  return data;
-}
